+++ Frontend/src/features/reviews/components/ReviewSlider.tsx
@@ -60,7 +60,7 @@ export default function ReviewSlider(){
     return (
         <section className="center-col gap-10 z-10">
 
-            <section ref={reviewsRef} className="w-[20rem] laptop:w-[61rem] flex shrink-0 overflow-hidden gap-2">
+            <section ref={reviewsRef} className="w-[20rem] laptop:w-[61rem] flex shrink-0 overflow-hidden gap-2 h-[13.1rem]">
                 {
                     reviews.map((review, index)=>(
                         <ReviewBox key={index} review={review}/>
@@ -68,8 +68,8 @@ export default function ReviewSlider(){
                 }
             </section>
             <div className="flex z-30 mb-10">
-                <IoIosArrowBack className="w-10 h-10 text-white cursor-pointer" onClick={()=>{scroll(false, "left",true)}}/>
-                <IoIosArrowForward className="w-10 h-10 text-white cursor-pointer" onClick={()=>{scroll(false, "right", true)}}/>
+                <IoIosArrowBack className="w-10 h-10 text-boxes cursor-pointer" onClick={()=>{scroll(false, "left",true)}}/>
+                <IoIosArrowForward className="w-10 h-10 text-boxes cursor-pointer" onClick={()=>{scroll(false, "right", true)}}/>
             </div>
 
         </section>