+++ Frontend/src/features/specialties/enums/specialities.enum.ts
@@ -8,3 +8,9 @@ export enum InformationTypeEnum{
     section = "section",
     selection = "selection"
 }
+
+export enum FormErrorEnum{
+    InconsistentDates = "La fecha de entrada es posterior a la de salida",
+    EmptyDate = "Ingrese todas las fechas",
+    NegativeSalary = "El salario debe ser mayor que 0",
+}