+++ Frontend/src/features/specialties/interfaces/specialties.interface.ts
@@ -16,3 +16,36 @@ export interface IOption{
     imageName: string
     
 }
+
+export interface ICompensationResponse{
+    antiquity: number
+    dayMonth: number 
+    monthIntegration: number
+    proportionalSac: number
+    holiday: number
+    noticedSust: number
+    noticedSac: number
+    total: number
+}
+
+export interface IFormError{
+    dateError: string,
+    salaryError: string
+}
+
+export interface ICompensationRequest{
+    dates: IDates
+    status: IStatus
+}
+
+interface IDates{
+    initial_date: string
+    final_date: string
+
+}
+interface IStatus{
+    salary: number 
+    reason: string
+    noticed: boolean
+    black: boolean
+}